+++ src/playchess/scorer.py
@@ -8,8 +8,8 @@ PIECE_SCORES = {Piece.WHITE_KING: 0, Piece.BLACK_KING: 0,
                 Piece.WHITE_BISHOP: 3, Piece.BLACK_BISHOP: -3,
                 Piece.WHITE_KNIGHT: 3, Piece.BLACK_KNIGHT: -3,
                 Piece.WHITE_PAWN: 1, Piece.BLACK_PAWN: -1}
-CHECKMATE_ABS_SCORE = 1000
-STALEMATE_SCORE = 0
+CHECKMATE_ABS_SCORE = 1000.0
+STALEMATE_SCORE = 0.0
 
 POSITIONAL_SCORES = {Piece.WHITE_KNIGHT: ((1, 1, 1, 1, 1, 1, 1, 1),
                                           (1, 2, 2, 2, 2, 2, 2, 1),
@@ -94,7 +94,7 @@ POSITIONAL_SCORES = {Piece.WHITE_KNIGHT: ((1, 1, 1, 1, 1, 1, 1, 1),
                      }
 
 
-def find_game_score(game: Game, use_positional_scoring: bool) -> int:
+def find_game_score(game: Game, use_positional_scoring: bool) -> float:
     """Finds the score of a chess game based on the material."""
 
     if game.turn.is_white() and game.check_mate:
@@ -104,16 +104,16 @@ def find_game_score(game: Game, use_positional_scoring: bool) -> int:
     elif game.stale_mate:
         return STALEMATE_SCORE
 
-    score = 0
+    score = 0.0
     for row_num, row in enumerate(game.chess_board):
         for col_num, piece_ in enumerate(row):
             if piece_.is_none():
                 continue
             positional_score_matrix = POSITIONAL_SCORES.get(piece_, None)
-            positional_score = 0
+            positional_score = 0.0
             if positional_score_matrix and use_positional_scoring:
                 positional_score = positional_score_matrix[row_num][col_num]
 
-            score = score + PIECE_SCORES[piece_] + (positional_score * 0.1)
+            score = score + float(PIECE_SCORES[piece_]) + (positional_score * 0.1)
 
     return score
